+++ test_cpal.rs
@@ -1,3 +1,0 @@
-use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
-use cpal::Sample;
-fn main() {}