+++ test_sample.rs
@@ -1,3 +1,0 @@
-use cpal::Sample;
-fn test_it<T: Sample>(val: T) -> f32 { val.to_f32() }
-fn main() {}